Changed paths: M       CHANGES.txt M       cat12.m M       cat_amap.mexmaci64 M       cat_io_cprintf.m M       cat_io_csv.m M       cat_io_matlabversion.m M       cat_io_report.m M       cat_main_reportfig.m M       cat_ornlm.mexmaci64 M       cat_run_newcatch.m M       cat_sanlm.mexmaci64 M       cat_update.m M       cat_vbdist.mexmaci64 M       cat_vol_downcut.mexmaci64 M       cat_vol_eidist.c M       cat_vol_eidist.mex M       cat_vol_eidist.mexmaci64 M       cat_vol_genus0.mexmaci64 M       cat_vol_gradient3.mexmaci64 M       cat_vol_interp3f.mexmaci64 M       cat_vol_laplace3.mexmaci64 M       cat_vol_laplace3R.mexmaci64 M       cat_vol_localstat.c M       cat_vol_localstat.mex M       cat_vol_localstat.mexmaci64 M       cat_vol_median3.mexmaci64 M       cat_vol_median3c.mexmaci64 M       cat_vol_pbtp.cpp MM      cat_vol_pbtp.m M       cat_vol_pbtp.mex M       cat_vol_pbtp.mexmaci64 MM      cat_vol_simgrow.m M       cat_vol_simgrow.mexmaci64 M       compile.m M       internal/cat_vol_cMRegularizarNLM3D.mexmaci64
(1) Changed: Made some Matlab code compatibe to Octave (Thanks to Guillaume
    Flandin for suggestions).
(2) Fixed: Corrected output parameters of cat_vol_pbtp.cpp.
(3) Changed: Added timeout parameter to urlread to prevent long wating times
    if server is not responding.


git-svn-id: svn://141.35.69.214/cat12@1524 9f44b166-36c4-483c-8295-4af200601734

+++ cat_vol_eidist.c
@@ -280,7 +280,7 @@ void mexFunction(int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[]) {
    */
   if ( verb ) {
     printf("\ncat_vol_eidist.c debuging mode:\n  Initialize Parameter: \n");
-    printf("    size(B) = %d %d %d\n",sL[0],sL[1],sL[2]); 
+    printf("    size(B) = %d %d %d\n",(int)sL[0],(int)sL[1],(int)sL[2]); 
     printf("    vx_vol  = %0.0f %0.0f %0.0f\n",s1,s2,s3); 
     printf("    euclid  = %d\n",(int) euclid); 
     printf("    setnan  = %d\n",(int) setnan);
